Created event dispatcher implementation
Event processing must be done on the GUI thread, and therefore must be
initially processed by the ServerEvent's slot. The Eventdispatcher will
connect to the eventSignal, and will process any event that emitted on
the GUI thread. The hope is to keep event processing in one place and
to aviod the if this/else that madness.

Anything that cares about event processing, only has to 'connect' to the
dispatcher instance specifying the type of event it is interested in, and
a method with a signature specifying the event properties that it cares
about.

For example a class might define a method, and connect to the dispatcher
as follows:

def process_chat_msg(self, message):
    # process message

dispatcher.connect(EventType.CHAT, self.process_chat_msg)

There is a downside to this approach in that, if there is a change to the
name of an Event's field, this will break any methods defined by any items
connected to the dispatcher. Proper unit testing can help with this though.

NOTE: I've also introduced a test framework (requires python-nose).

You can run the tests from the client/qt folder in two ways:
- nosetests
- make nose