Add dealt cards  to the cardbox.
Also sort them using the non-trump sorting. The trump sorting still
needs to be done. The sorting is done in the cardbox instead of in
the actual card widgets because the card widgets don't know what
suit is trump.

Card suits were reorderd to make them easier to sort to alternating
colours.