Refactor play_bella to not require card.
The game can figure out the card because there's only one that is valid
to play when this command is called. This avoids the client having to
first select the card.

Tested with bots and they seem to have someone get bella one out of
every four hands. Once someone gets the first of bells the chance
they'll get the last of bells is 25%, so this seems right.