Start of a mochiweb server.
Redirects to lobby after looking up the client. Then it doesn't actually reply
yet, but it prints if it finds the client by cookie.

The tarabish server cookie and the browser "session" cookie are set to the same
value, only the browser one is a string so needs to be converted.