Version 2 of the protocol.
Pass the client cookie for calls so the socket doesn't need to be kept
open for a client to stay logged in. Currently if they don't quit()
before disconnecting they will never be logged out, so this will need
to be fixed.

As login isn't required the two services could be combined. Two
connections will still be required as the get_events() calls are
blocking.